Refactor: Commands to their specific files
Commands moved to files with respective command names for ease of maintenance.
