+++ timezones_cli/commands/add.py
@@ -1,0 +1,44 @@
+"""
+Sub command to add timezones for quick glance.
+"""
+
+import click
+from typing import List
+
+from timezones_cli.utils import (
+    variables,
+    console,
+    validate_timezone,
+    check_config,
+)
+
+
+@click.command()
+@click.argument("timezone")
+def add(timezone: str):
+    """
+    Add timezone to the config file.
+    """
+    validate_timezone(timezone)
+    config_file = variables.config_file
+
+    if not check_config():
+        with open(config_file, "w+") as newfile:
+            pass
+
+    with open(config_file, "r+") as file:
+        data: List = [line.rstrip() for line in file]
+
+        if timezone in data:
+            return console.print(
+                f"[bold green]Timezone already exists:[/bold green] [bold red]{timezone}:x:[/bold red]"
+            )
+
+        # Read file
+        file.read()
+
+        # Add to the end of the file.
+        file.write(f"{timezone}\n")
+        console.print(
+            f"[bold green]New timezone added successfully:[/bold green] [bold blue]{timezone}[/bold blue] :white_check_mark:"
+        )
